+++ app/src/main/java/com/rashwan/reactive_popular_movies/service/MoviesService.java
@@ -13,6 +13,6 @@ import rx.Observable;
 public interface MoviesService {
      Observable<MoviesResponse> getPopularMovies(int page);
      Observable<MoviesResponse> getTopRatedMovies(int page);
-     Observable<TrailersResponse> getMovieTrailers(int id);
-     Observable<ReviewResponse> getMovieReview(int id);
+     Observable<TrailersResponse> getMovieTrailers(long id);
+     Observable<ReviewResponse> getMovieReview(long id);
 }
